fix: Remove unused backends to flatten dependencies
`Pancurses` and `Ncurses` (default features) are only used if crossterm fails. As this is very unlikely, because crossterm
support all unix terminals and windows terminals down to win7, I deactivated these default features. As a result the user don't
have to install the ncurses library on the system first, to be able to build the commit-tool.